Fix warning: Implicit conversion loses integer precision

+++ SocketRocket/SRWebSocket.m
@@ -1349,7 +1349,7 @@ static const size_t SRFrameHeaderOverhead = 32;
     uint8_t *frameBuffer = (uint8_t *)frameData.mutableBytes;
 
     // set fin
-    frameBuffer[0] = SRFinMask | opCode;
+    frameBuffer[0] = SRFinMask | (uint8_t)opCode;
 
     // set the mask and header
     frameBuffer[1] |= SRMaskMask;
